pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fosm-mlops"
version = "0.1.0"
description = "Fiber-optic pipeline monitoring ML/MLOps project"
readme = "README.md"
authors = [{name = "FiberOps Team", email = "mlops@example.com"}]
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.23",
    "pandas>=1.5",
    "scipy>=1.9",
    "pyarrow>=12.0",
    "scikit-learn>=1.1",
    "imbalanced-learn>=0.10",
    "xgboost>=1.7",
    "lightgbm>=3.3",
    "mlflow>=2.7",
    "hydra-core>=1.3",
    "pydantic>=1.10",
    "fastapi>=0.103",
    "uvicorn>=0.23",
    "structlog>=23.1",
    "great-expectations>=0.16",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.0",
    "tensorflow>=2.11",
    "PyYAML>=6.0",
    "joblib>=1.2",
    "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "ruff>=0.0.292",
    "black>=23.9",
    "mypy>=1.6",
    "types-PyYAML",
    "types-requests",
    "pre-commit>=3.4",
    "jupyterlab>=4.0",
    "notebook>=7.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]
ignore = ["E203", "E266", "E501"]

[tool.mypy]
python_version = "3.10"
plugins = []
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --cov=fosm_mlops --cov-report=term-missing"
testpaths = ["tests"]

[project.scripts]
fosm-train = "fosm_mlops.pipelines.train_pipeline:main"
